Changing file name and fixing all the bugs

+++ secure_admin_run.py
@@ -1,6 +1,4 @@
-from cmath import e
 import os
-import sys
 import time
 import datetime
 import subprocess
@@ -23,11 +21,11 @@ def ban_ip_thread(jail_name, bantime, maxretry, failurewindow, log_path, failure
     ip_start_time_dict = {}
     ip_end_time_dict = {}
 
+    # print("[*] Starting multiprocess with id: " + str(os.getpid()))
+
     f = subprocess.Popen(['tail','-F', '-n', '1', log_path],\
             stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     while True:
-        if kill():
-            break
         line = f.stdout.readline().decode('utf-8')
         # Check for regex match for failed login attempts
         if failure_regex.search(line):
@@ -73,86 +71,116 @@ def ban_ip_thread(jail_name, bantime, maxretry, failurewindow, log_path, failure
 
 if __name__ == '__main__':
     try :
+        config_file = 'custom_jail.conf'
         while True:
             cp = ConfigParser.RawConfigParser()
-            cp.read('custom_jail.conf')
+            cp.read(config_file)
             jails = cp.sections()
             for jail in jails:
-                print(jail)
-                bantime = cp.getint(jail, 'bantime')
-                print(bantime)
-                maxretry = cp.getint(jail, 'maxretry')
-                print(maxretry)
-                failurewindow = cp.getint(jail, 'failurewindow')
-                print(failurewindow)
+                value_changed = False
+
+                # Track if the value is changed in the config file
+                if cp.get(jail, 'bantime') != cp.get(jail, 'bantime_old'):
+                    cp.set(jail, 'bantime_old', int(cp.get(jail, 'bantime')))
+                    value_changed = True
+                bantime = int(cp.get(jail, 'bantime'))
+
+                if cp.get(jail, 'maxretry') != cp.get(jail, 'maxretry_old'):
+                    cp.set(jail, 'maxretry_old', int(cp.get(jail, 'maxretry')))
+                    value_changed = True
+                maxretry = int(cp.get(jail, 'maxretry'))
+
+                if cp.get(jail, 'failurewindow') != cp.get(jail, 'failurewindow_old'):
+                    cp.set(jail, 'failurewindow_old', int(cp.get(jail, 'failurewindow')))
+                    value_changed = True
+                failurewindow = int(cp.get(jail, 'failurewindow'))
+
                 failure_regex = cp.get(jail, 'regex')
-                print(failure_regex)
+
                 logpath = cp.get(jail, 'logpath')
-                print(logpath)
+
                 enabled = cp.getboolean(jail, 'enabled')
-                print(enabled)
 
                 # Start/Stop the thread for each jail
                 if jail == 'ssh':
                     if enabled:
                         print('[+] SSH jail enabled')
-                        if 'ssh_jail' in locals():
-                            if ssh_jail.is_alive():
+                        if 'ssh_jail' in locals() and ssh_jail.is_alive():
                                 print('[+] SSH jail already running')
+                                if value_changed:
+                                    #Stop the thread and start a new one
+                                    print('[+] Restarting SSH jail thread')
+                                    ssh_jail.terminate()
+                                    ssh_jail = multiprocessing.Process(target=ban_ip_thread, args=('ssh', bantime, maxretry, failurewindow, logpath, failure_regex))
+                                    ssh_jail.start()
                         else:
                             ssh_jail = multiprocessing.Process(target=ban_ip_thread, args=('ssh', bantime, maxretry, failurewindow, logpath, failure_regex))
                             ssh_jail.start()
                     else:
                         print('[-] SSH jail disabled')
                         # Kill the ssh thread if its running
-                        if 'ssh_jail' in locals():
-                            if ssh_jail.is_alive():
+                        if 'ssh_jail' in locals() and ssh_jail.is_alive():
                                 ssh_jail.terminate()
                 if jail == 'wordpress':
                     if enabled:
                         print('[+] Wordpress jail enabled')
-                        if 'wordpress_jail' in locals():
-                            if wordpress_jail.is_alive():
+                        if 'wordpress_jail' in locals() and wordpress_jail.is_alive():
                                 print('[+] Wordpress jail already running')
+                                if value_changed:
+                                    #Stop the thread and start a new one
+                                    print('[+] Restarting Wordpress jail thread')
+                                    wordpress_jail.terminate()
+                                    wordpress_jail = multiprocessing.Process(target=ban_ip_thread, args=('wordpress', bantime, maxretry, failurewindow, logpath, failure_regex))
+                                    wordpress_jail.start()
                         else:
                             wordpress_jail = multiprocessing.Process(target=ban_ip_thread, args=('wordpress', bantime, maxretry, failurewindow, logpath, failure_regex))
                             wordpress_jail.start()
                     else:
                         print('[-] Wordpress jail disabled')
                         # Kill the wordpress thread if its running
-                        if 'wordpress_jail' in locals():
-                            if wordpress_jail.is_alive():
+                        if 'wordpress_jail' in locals() and wordpress_jail.is_alive():
                                 wordpress_jail.terminate()
                 if jail == 'phpmyadmin':
                     if enabled:
                         print('[+] Phpmyadmin jail enabled')
-                        if 'phpmyadmin_jail' in locals():
-                            if phpmyadmin_jail.is_alive():
+                        if 'phpmyadmin_jail' in locals() and phpmyadmin_jail.is_alive():
                                 print('[+] Phpmyadmin jail already running')
+                                if value_changed:
+                                    #Stop the thread and start a new one
+                                    print('[+] Restarting Phpmyadmin jail thread')
+                                    phpmyadmin_jail.terminate()
+                                    phpmyadmin_jail = multiprocessing.Process(target=ban_ip_thread, args=('phpmyadmin', bantime, maxretry, failurewindow, logpath, failure_regex))
+                                    phpmyadmin_jail.start()
                         else:
                             phpmyadmin_jail = multiprocessing.Process(target=ban_ip_thread, args=('phpmyadmin', bantime, maxretry, failurewindow, logpath, failure_regex))
                             phpmyadmin_jail.start()
                     else:
                         print('[-] Phpmyadmin jail disabled')
                         # Kill the phpmyadmin thread if its running
-                        if 'phpmyadmin_jail' in locals():
-                            if phpmyadmin_jail.is_alive():
+                        if 'phpmyadmin_jail' in locals() and phpmyadmin_jail.is_alive():
                                 phpmyadmin_jail.terminate()
                 if jail == 'joomla':
                     if enabled:
                         print('[+] Joomla jail enabled')
-                        if 'joomla_jail' in locals():
-                            if joomla_jail.is_alive():
+                        if 'joomla_jail' in locals() and joomla_jail.is_alive():
                                 print('[+] Joomla jail already running')
+                                if value_changed:
+                                    #Stop the thread and start a new one
+                                    print('[+] Restarting Joomla jail thread')
+                                    joomla_jail.terminate()
+                                    joomla_jail = multiprocessing.Process(target=ban_ip_thread, args=('joomla', bantime, maxretry, failurewindow, logpath, failure_regex))
+                                    joomla_jail.start()
                         else:
                             joomla_jail = multiprocessing.Process(target=ban_ip_thread, args=('joomla', bantime, maxretry, failurewindow, logpath, failure_regex))
                             joomla_jail.start()
                     else:
                         print('[-] Joomla jail disabled')
                         # Kill the joomla thread if its running
-                        if 'joomla_jail' in locals():
-                            if joomla_jail.is_alive():
+                        if 'joomla_jail' in locals() and joomla_jail.is_alive():
                                 joomla_jail.terminate()
+                if value_changed:
+                        with open(config_file, 'w') as configfile:
+                            cp.write(configfile)
             time.sleep(2)
     except Exception as e:
         print(e.args)
